Criando objetoColunasDinamicas e Refatorando, deixando TabelaDinamica stateless
