+++ src/componentes/dres/PrestacaoDeContas/ListaPrestacaoDeContas/objetoColunasDinamicas.js
@@ -1,0 +1,30 @@
+export const colunasNaoRecebidas = [
+    {field: 'unidade_eol', header: 'Código Eol'},
+    {field: 'unidade_nome', header: 'Nome da unidade'},
+    {field: 'tecnico_responsavel', header: 'Técnico atribuído'},
+    {field: 'data_recebimento', header: 'Data de recebimento'},
+    {field: 'status', header: 'Status'},
+    {field: 'acoes', header: 'Ações'},
+];
+
+export const colunasEmAnalise = [
+    {field: 'unidade_eol', header: 'Código Eol'},
+    {field: 'unidade_nome', header: 'Nome da escola'},
+    {field: 'processo_sei', header: 'Processo SEI'},
+    {field: 'data_recebimento', header: 'Data de recebimento'},
+    {field: 'data_ultima_analise', header: 'Última análise'},
+    {field: 'tecnico_responsavel', header: 'Técnico atribuído'},
+    {field: 'acoes', header: 'Ações'},
+];
+
+export const colunasAprovada = [
+    {field: 'unidade_eol', header: 'Código Eol'},
+    {field: 'unidade_nome', header: 'Nome da escola'},
+    {field: 'processo_sei', header: 'Processo SEI'},
+    {field: 'data_recebimento', header: 'Data de recebimento'},
+    {field: 'data_ultima_analise', header: 'Última análise'},
+    {field: 'tecnico_responsavel', header: 'Quem está validando'},
+    {field: 'tecnico_responsavel', header: 'Devolução ao tesouro'},
+    {field: 'status', header: 'Status'},
+    {field: 'acoes', header: 'Ações'},
+];
